+++ apps/include/proto.h
@@ -38,7 +38,6 @@ int parseframe(uint8_t b, struct frame_header *hdr);
 void writebyte(uint8_t b);
 void write(uint8_t *buf, size_t nbytes);
 uint8_t readbyte();
-uint8_t readbyte_ledflash(int ledvalue, int loopcount);
 void read(uint8_t *buf, size_t nbytes);
 
 #endif